feat: save parallel
